requirements.txt
pydantic==2.8.2
python-telegram-bot==21.3
requests==2.32.3
python-dotenv==1.0.1
google-api-python-client==2.137.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
pymongo==4.8.0
celery==5.4.0
redis==5.0.7
async-timeout==4.0.3